requirements.txt
requests==2.31.0
python-dotenv==1.0.1
groq==0.5.0